requirements.txt
asgiref==3.6.0
async-generator==1.10
atomicwrites==1.4.1
attrs==22.2.0
autopep8==2.0.2
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudinary==1.32.0
colorama==0.4.6
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
cryptography==39.0.1
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==3.0.0
Django==4.1.6
django-allauth==0.52.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-extensions==3.2.1
django-filter==22.1
django-guardian==2.4.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-nine==0.2.7
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-guardian==0.3.0
djangorestframework-simplejwt==5.2.2
docstring-parser==0.14.1
drf-action-serializer==1.0.5
drf-nested-routers==0.93.4
drf-spectacular==0.25.1
drf-standardized-errors==0.12.5
drf-yasg==1.21.5
elasticsearch==7.17.9
elasticsearch-dsl==7.4.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==17.0.0
filelock==3.9.0
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
msgpack==1.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
pipenv==2022.12.19
platformdirs==2.6.2
pluggy==1.0.0
py==1.11.0
py-algorand-sdk==1.17.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==2.6.0
PyNaCl==1.5.0
pyrsistent==0.19.3
PySocks==1.7.1
pyteal==0.18.1
pytest==6.2.5
pytest-base-url==2.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-selenium==4.0.0
pytest-variables==2.0.0
python-dateutil==2.8.2
python-decouple==3.7
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.8.2
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tenacity==6.3.1
toml==0.10.2
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.17.1
virtualenv-clone==0.5.7
Werkzeug==2.2.3
wsproto==1.2.0
